requirements.txt
asgiref==3.7.2
daphne==4.0.0
dj-database-url==2.1.0
Django==4.2.4
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.2
djangorestframework==3.14.0
dotenv==0.0.5
frozenlist==1.4.0
gunicorn==21.2.0
psycopg2-binary==2.9.7
tzdata==2023.3

